Add files via upload
Logo ideas for openSUSECon 2024. I know yellow isn't part of the guidelines but it adds some brightness to it. If not, one of the two files uses all colors and tints from the guidelines. I also tried to keep the round style used on the community summit logo, but it can be any other shape. Let me know what you think.